🏗️ Combine pipeline start and webhook entrypoint scripts
Having two entrypoints proved problematic for continuing the pipeline;
it is easier to always have one point of entry.

The start script parses both command line arguments and input payload,
and uses the one where data is available.

The parameters are always passed on as metadata to the next node.